Split the diff cover and diff quality main files
This is generally cleaner as the two done actually share much logic

This will also benefit users who would rather use the python command to run these tools rather than the provided entry points
