+++ diff_cover/__init__.py
@@ -1,3 +1,5 @@
+from __future__ import unicode_literals
+
 VERSION = '1.0.7'
 DESCRIPTION = 'Automatically find diff lines that need test coverage.'
 QUALITY_DESCRIPTION = 'Automatically find diff lines with quality violations.'
