+++ diff_cover/command_runner.py
@@ -1,4 +1,3 @@
-import re
 import six
 import subprocess
